% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_example_data.R
\name{download_testdata_IceChaRt}
\alias{download_testdata_IceChaRt}
\title{Download example raster data for IceChaRt}
\usage{
download_testdata_IceChaRt(dest = system.file("extdata", package = "IceChaRt"))
}
\arguments{
\item{dest}{Directory where the files should be saved.
Defaults to the \code{extdata} folder of the installed IceChaRt package.}
}
\value{
Invisibly returns the path where files were saved.
}
\description{
Downloads bigger test data for the IceChaRt package to try out all functions.
Only needs to be run once.
}
\examples{
\dontrun{
download_testdata_IceChaRt()

# Ready to use:
co_pol    <- terra::rast(system.file("extdata", "s1_20201101_hh.tif", package = "IceChaRt"))
cross_pol <- terra::rast(system.file("extdata", "s1_20201101_hv.tif", package = "IceChaRt"))
}

}
