% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchCISfiles.R
\name{search_cis_icechart}
\alias{search_cis_icechart}
\title{List of available Ice Chart files of the Canadian Ice Service (CIS)}
\usage{
search_cis_icechart(region = "Eastern_Arctic", year = "2020")
}
\arguments{
\item{region}{"East_Coast", “Eastern_Arctic”, "Great_Lakes", “Hudson_Bay”, "Western_Arctic"}

\item{year}{Choose between "2006" to current year}
}
\value{
data.frame with dataname, URLs, date and version
}
\description{
This function searches for the weekly ice charts from the CIS
for a specific region and year, and prints the results to the users console.
}
\examples{
\dontrun{

search_cis_icechart(region="Eastern_Arctic", year="2020")
}

}
